Separate Extent and ExtentList
